requirements.txt
numpy>=1.18.5
numba>=0.51.2
pillow>=8.0.1
opencv-python>=4.2.0